add class DuplicateFilesReport
